+++ lib/async/service/supervisor.rb
@@ -12,6 +12,7 @@ require_relative "supervisor/client"
 
 require_relative "supervisor/memory_monitor"
 require_relative "supervisor/process_monitor"
+require_relative "supervisor/utilization_monitor"
 
 require_relative "supervisor/environment"
 require_relative "supervisor/supervised"